Upgrade RTS dependencies
