+++ rts/motoko-rts/Cargo.toml
@@ -35,7 +35,7 @@ classical_persistence = ["libc"]
 enhanced_orthogonal_persistence = ["incremental_gc"]
 
 [dependencies]
-libc = { version = "0.2.139", default_features = false, optional = true }
+libc = { version = "0.2.139", default-features = false, optional = true }
 motoko-rts-macros = { path = "../motoko-rts-macros" }
 
 [profile.dev]
