+++ rts/motoko-rts/src/lib.rs
@@ -4,12 +4,12 @@
 #![feature(
     arbitrary_self_types,
     core_intrinsics,
-    panic_info_message,
     proc_macro_hygiene,
     // // We do not need simd but this flag enables `core::arch:wasm64`.
     // // See https://github.com/rust-lang/rust/issues/90599
     simd_wasm64
 )]
+#![allow(internal_features)]
 
 // c.f. https://os.phil-opp.com/heap-allocation/#dynamic-memory
 extern crate alloc;
@@ -128,24 +128,17 @@ pub(crate) unsafe fn rts_trap_with(msg: &str) -> ! {
 #[panic_handler]
 fn panic(info: &core::panic::PanicInfo) -> ! {
     unsafe {
-        if let Some(msg) = info.payload().downcast_ref::<&str>() {
-            println!(1000, "RTS panic: {}", msg);
-        } else if let Some(args) = info.message() {
-            let mut buf = [0 as u8; 1000];
-            let mut fmt = print::WriteBuf::new(&mut buf);
-            let _ = core::fmt::write(&mut fmt, *args);
-            print::print(&fmt);
-        } else {
-            println!(1000, "RTS panic: weird payload");
-        }
-
+        let message = info.message();
         if let Some(location) = info.location() {
             println!(
                 1000,
-                "panic occurred in file '{}' at line {}",
+                "panic occurred in file '{}' at line {}: {}",
                 location.file(),
                 location.line(),
+                message,
             );
+        } else {
+            println!(1000, "RTS panic: {message}");
         }
 
         rts_trap_with("RTS panicked");
