+++ src/codegen/compile_classical.ml
@@ -2544,7 +2544,8 @@ module Closure = struct
     Tagged.load_forwarding_pointer env ^^
     Tagged.load_field env (funptr_field env) ^^
     (* All done: Call! *)
-    G.i (CallIndirect (nr ty)) ^^
+    let table_index = 0l in
+    G.i (CallIndirect (nr table_index, nr ty)) ^^
     FakeMultiVal.load env (Lib.List.make n_res I32Type)
 
   let static_closure env fi : int32 =
@@ -13015,7 +13016,7 @@ and conclude_module env set_serialization_globals start_fi_o =
   let emodule =
     let open Wasm_exts.CustomModule in
     { module_;
-      dylink = None;
+      dylink0 = [];
       name = { empty_name_section with function_names =
                  List.mapi (fun i (f,n,_) -> Int32.(add ni' (of_int i), n)) funcs;
                locals_names =
