+++ src/languageServer/.ocamlformat
@@ -1,4 +1,4 @@
-version=0.26.1
+version=0.26.2
 ocaml-version=4.12
 break-infix=fit-or-vertical
 doc-comments=before
