+++ test/bench/ok/alloc.drun-run.ok
@@ -2,7 +2,7 @@ ingress Completed: Reply: 0x4449444c016c01b3c4b1f204680100010a000000000000000001
 ingress Completed: Reply: 0x4449444c0000
 debug.print: (+268_435_456, 3_071_963_514)
 ingress Completed: Reply: 0x4449444c0000
-debug.print: (+268_435_456, 3_070_488_954)
+debug.print: (+268_435_456, 3_070_505_338)
 ingress Completed: Reply: 0x4449444c0000
-debug.print: (+268_435_456, 3_070_488_954)
+debug.print: (+268_435_456, 3_070_505_338)
 ingress Completed: Reply: 0x4449444c0000
