+++ test/bench/ok/heap-32.drun-run.ok
@@ -1,5 +1,5 @@
 ingress Completed: Reply: 0x4449444c016c01b3c4b1f204680100010a00000000000000000101
 ingress Completed: Reply: 0x4449444c0000
-debug.print: (50_227, +29_863_068, 769_836_717)
-debug.print: (50_070, +32_992_212, 831_339_788)
+debug.print: (50_227, +29_863_068, 739_048_937)
+debug.print: (50_070, +32_992_212, 803_212_043)
 ingress Completed: Reply: 0x4449444c0000
