+++ test/bench/ok/heap-64.drun-run.ok
@@ -1,5 +1,5 @@
 ingress Completed: Reply: 0x4449444c016c01b3c4b1f204680100010a00000000000000000101
 ingress Completed: Reply: 0x4449444c0000
-debug.print: (49_965, +47_942_744, 1_027_584_937)
-debug.print: (49_806, +47_960_000, 1_036_639_548)
+debug.print: (49_965, +47_942_744, 997_700_939)
+debug.print: (49_806, +47_960_000, 1_006_736_620)
 ingress Completed: Reply: 0x4449444c0000
