+++ test/bench/ok/palindrome.drun-run.ok
@@ -1,9 +1,9 @@
 ingress Completed: Reply: 0x4449444c016c01b3c4b1f204680100010a00000000000000000101
 ingress Completed: Reply: 0x4449444c0000
-debug.print: (true, +1_188, 13_154)
-debug.print: (false, +1_188, 12_203)
-debug.print: (false, +1_188, 13_133)
-debug.print: (true, +868, 13_262)
-debug.print: (false, +868, 11_725)
-debug.print: (false, +868, 13_633)
+debug.print: (true, +1_188, 13_158)
+debug.print: (false, +1_188, 12_207)
+debug.print: (false, +1_188, 13_137)
+debug.print: (true, +868, 13_266)
+debug.print: (false, +868, 11_729)
+debug.print: (false, +868, 13_214)
 ingress Completed: Reply: 0x4449444c0000
