+++ test/bench/ok/region-mem.drun-run.ok
@@ -1,4 +1,4 @@
 ingress Completed: Reply: 0x4449444c016c01b3c4b1f204680100010a00000000000000000101
 ingress Completed: Reply: 0x4449444c0000
-debug.print: {heap_diff = 0; instr_diff = 5_964_300_649}
+debug.print: {heap_diff = 0; instr_diff = 5_939_134_825}
 ingress Completed: Reply: 0x4449444c0000
