+++ test/bench/ok/region0-mem.drun-run.ok
@@ -1,4 +1,4 @@
 ingress Completed: Reply: 0x4449444c016c01b3c4b1f204680100010a00000000000000000101
 ingress Completed: Reply: 0x4449444c0000
-debug.print: {heap_diff = 0; instr_diff = 6_392_119_657}
+debug.print: {heap_diff = 0; instr_diff = 6_366_953_833}
 ingress Completed: Reply: 0x4449444c0000
