+++ test/ld/ok/representative.linked.wat.ok
@@ -11,8 +11,7 @@
     call $not_yet_imported1
     call $not_yet_imported2)
   (func $resolved_export (type 0))
-  (func $__wasm_call_ctors (type 0)
-    call $__wasm_apply_data_relocs)
+  (func $__wasm_call_ctors (type 0))
   (func $__wasm_apply_data_relocs (type 0))
   (func $square (type 1) (param i32) (result i32)
     call $resolved_export
@@ -22,7 +21,10 @@
     i32.mul)
   (func $link_start (type 0)
     call $__wasm_call_ctors)
+  (func $link_start.1 (type 0)
+    call $__wasm_apply_data_relocs
+    call $link_start)
   (table (;0;) 0 0 funcref)
   (memory (;0;) i64 2)
   (global (;0;) i64 (i64.const 65536))
-  (start $link_start))
+  (start $link_start.1))
