+++ test/lsp-int/lsp-int.cabal
@@ -14,8 +14,8 @@ executable lsp-int
   main-is:             Main.hs
   -- other-modules:
   other-extensions:    OverloadedStrings, DuplicateRecordFields
-  build-depends:       base ^>=4.17
-                     , text ^>=2.0
+  build-depends:       base
+                     , text
                      , hspec
                      , HUnit
                      , filepath
