+++ test/random/qc-motoko.cabal
@@ -15,8 +15,8 @@ executable qc-motoko
   other-modules:       Embedder, Turtle.Pipe
   other-extensions:    ConstraintKinds, StandaloneDeriving, DataKinds, KindSignatures, GADTs, MultiParamTypeClasses
   ghc-options:         -O -threaded -with-rtsopts=-N2
-  build-depends:       base ^>=4.17
-                     , text ^>=2.0
+  build-depends:       base
+                     , text
                      , process
                      , exceptions
                      , managed
