--- test/run/ok/float-fmt-classical.run.ok
@@ -1,32 +1,0 @@
-fixed Float
-3
-3.1
-3.1416
-3.141592654
-3.1415926535897931
-3.14159265358979312
-3.14159265358979311600
-exponential Float
-3e+00
-3.1e+00
-3.1416e+00
-3.141592654e+00
-3.1415926535897931e+00
-3.14159265358979312e+00
-3.14159265358979311600e+00
-generic Float
-3
-3
-3.142
-3.14159265
-3.141592653589793
-3.1415926535897931
-3.141592653589793116
-hex Float
-0x2p+1
-0x1.9p+1
-0x1.9220p+1
-0x1.921fb5444p+1
-0x1.921fb54442d18000p+1
-0x1.921fb54442d180000p+1
-0x1.921fb54442d180000000p+1
