+++ test/run/ok/float-fmt.run-ir.ok
