+++ test/run/ok/float-fmt.run-low.ok
