+++ test/run/ok/float-fmt.run.ok
